# sanitize structured metadata values during query time in pipeline code