In an earlier PR ([here](https://github.com/grafana/loki/pull/13983)) we made a change to sanitize structured metadata names of disallowed characters, since when SM is added to pipeline stages in a users query the SM is turned into labels and in turned passed to Prometheus' label parser. That then results in an error if there's disallowed characters.

What we missed at the time was the same sanitization of structured metadata values. This PR sanitizes the value in the same way, replacing the invalid character with  a space ` `. While this likely requires users to query their structured metadata with a regex filter, it's at least a workaround. Perhaps we can find a way to return a warning to the user that their query hit SM that had invalid characters, and therefore they should consider changing their `MatchEqual` to `MatchRegex` if they're not seeing results.

This might also be a good time to revisit the topic of sanitizing or rejecting SM at ingestion time.